Refactor Indeed integration API
