--- lib/indeed/__init__.py
@@ -1,7 +1,0 @@
-# HTK Imports
-from htk.lib.indeed.api import IndeedDispositionSyncAPI
-
-
-__all__ = [
-    'IndeedDispositionSyncAPI'
-]
