+++ lib/indeed/api/__init__.py
@@ -1,0 +1,7 @@
+# HTK Imports
+from htk.lib.indeed.api.disposition_sync import IndeedDispositionSyncAPI
+
+
+__all__ = [
+    'IndeedDispositionSyncAPI',
+]
