+++ lib/indeed/constants.py
@@ -1,2 +1,2 @@
-INDEED_DISPOSITION_AUTH_TOKEN_URL = 'https://apis.indeed.com/oauth/v2/tokens'
-INDEED_DISPOSITION_GRAPHQL_API_URL = 'https://apis.indeed.com/graphql'
+INDEED_OAUTH_TOKEN_URL = 'https://apis.indeed.com/oauth/v2/tokens'
+INDEED_GRAPHQL_API_URL = 'https://apis.indeed.com/graphql'
